requirements.txt
flask
openai-whisper
scipy
